Additional Player forces can be added without hindering player movement.
Additional forces can be added to the player while they have constrained movement.
Essentially, if the player is moving faster than their max speed, they will be able to slow down and turn with inputs but not be able to continue speed.
Additional forces can be added in and the player will still carry that momentum. This is the root of how movement abilities will work.

There is a the issue of the Drag on the player always being too strong. This may need to be adjusted during gameplay or maybe rethought. Sliding mechanic can be achieved by reducing this number.