Sloped Movement implemented.
The player will move with consideration to the ground normal.

There are still a few issues.
If the player is moving too fast over a bump or cliff, they will keep their velocity momentum and become airborne.

When the player is moving on a slope and jumps, vertical velocity is directly added to the falling velocity and causes the player to super jump.