Added wall climbing implementation// BUGFIX. No Bouncing.
Wall climbing is implemented, but there's no actual input for it yet.

BUG Fix.
The player doesn bounce after wall jumps any more.
Velocities were incorrectly applied when jumping not considering the dumb hacky falling velocity the player has. This caused the velocity to buil up far higher than it should have and when the player touched the ground all the velocity gained was directly added to falling velocity causing the player to launch to the skies.