Changed How inputs were working
Jumping and wall running and climbing are now based on a "Rising Input".
This is currently the space key, when the input is pressed the player will do the corresponding input based on the current parameters. ie. if next to a wall and moving towards it. climb the wall. If no walls around, simply jump.